@@ -20,25 +20,24 @@
   },
   "homepage": "https://github.com/erdtman/solar-power-skir#readme",
   "dependencies": {
-    "axios": "^0.21.1",
-    "chart.js": "^2.8.0",
-    "dotenv": "^7.0.0",
-    "express": "^4.16.4",
-    "moment": "^2.24.0",
-    "moment-timezone": "^0.5.28",
-    "mongodb": "^3.6.3",
-    "vue": "^2.6.10",
-    "vue-chartjs": "^3.4.2",
-    "vue-router": "^3.0.2",
-    "vue-timers": "^2.0.2",
-    "vue2-touch-events": "^2.1.0"
+    "axios": "^0.26.1",
+    "chart.js": "^3.7.1",
+    "dotenv": "^16.0.0",
+    "express": "^4.17.3",
+    "moment": "^2.29.3",
+    "moment-timezone": "^0.5.34",
+    "mongodb": "^4.5.0",
+    "vue": "^3.2.33",
+    "vue-router": "^4.0.14",
+    "vue-timers": "^2.0.4",
+    "vue2-touch-events": "^3.2.2"
   },
   "devDependencies": {
-    "css-loader": "^2.1.1",
-    "vue-loader": "^15.7.0",
-    "vue-style-loader": "^4.1.2",
-    "vue-template-compiler": "^2.6.10",
-    "webpack": "^4.29.6",
-    "webpack-cli": "^3.3.6"
+    "css-loader": "^6.7.1",
+    "vue-loader": "^17.0.0",
+    "vue-style-loader": "^4.1.3",
+    "vue-template-compiler": "^2.6.14",
+    "webpack": "^5.72.0",
+    "webpack-cli": "^4.9.2"
   }
 }
