@@ -1,7 +1,5 @@
-import Vue from 'vue';
+import { createApp } from 'vue';
 import App from './App.vue';
 
-new Vue({
-  el: '#app',
-  render: h => h(App),
-});
+const app = createApp(App);
+app.mount('#app');
